Refactors naming of api files
