+++ flaskr/api/appliance_depot.py
